add in a maroc stand alone readout for testing
- also fix type of iFlag

+++ ssp_maroc_rol_include.c
@@ -42,7 +42,7 @@ sspMaroc_Download()
 void
 sspMaroc_Prestart()
 {
-  unsigned short iFlag;
+  unsigned int iFlag;
   int id, SSP_SLOT;
 
   /*****************
